+++ tests/models.py
@@ -33,7 +33,10 @@ class Article(models.Model):
     title = models.CharField(max_length=100)
     author = models.ForeignKey(Author, on_delete=models.CASCADE)
     publisher = models.ForeignKey(PeriodicalPublisher, related_name='published', on_delete=models.CASCADE)
-    release = models.DateTimeField(auto_now_add=True)
+    release = models.DateField()
+
+    class Meta:
+        get_latest_by = 'release'
 
     def __str__(self):
         return "%s by %s" % (self.title, self.author)
@@ -49,8 +52,11 @@ class Book(models.Model):
     title = models.CharField(max_length=50)
     author = models.ForeignKey(Author, on_delete=models.CASCADE)
     publisher = models.ForeignKey(Publisher, related_name='published', on_delete=models.CASCADE)
-    release = models.DateTimeField(auto_now_add=True)
+    release = models.DateField()
     pages = models.PositiveSmallIntegerField()
+
+    class Meta:
+        get_latest_by = 'release'
 
     def __str__(self):
         return "%s by %s" % (self.title, self.author)